refactor(tui): separate terminal palette concerns
