+++ station/src/station/view/dashboard.golden.test.tsx
@@ -26,8 +26,8 @@ import {
   StationThemeProvider,
   type StationTheme,
 } from "../../theme/index.js";
-import { parseStationTerminalPaletteObservation } from "../../theme/terminalPaletteObservation.js";
-import { lightTerminalColors } from "../../theme/test/terminalPaletteFixtures.js";
+import { parseStationTerminalPaletteObservation } from "../../theme/terminalPalette/observation.js";
+import { lightTerminalColors } from "../../theme/terminalPalette/test/fixtures.js";
 import { StationHoverProvider, StationMouseProvider } from "./stationMouseContext.js";
 
 function spanHex(span: ReturnType<typeof spanAtFrameCell>): string | undefined {
